pyproject.toml
[tool.poetry]
name = "action-poetry-update"
version = "0.1.0"
description = "GitHub action for updating Python dependencies"
authors = ["moneymeets GmbH <service@moneymeets.com>"]

packages = [
    { include = "actions_helper" }
]

[tool.poetry.dependencies]
python = "~3.12"

click="*"
jinja2 = "*"
pygithub = "*"

[tool.poetry.dev-dependencies]
ruff = "*"
pytest = "*"
pytest-cov = "*"


[tool.poetry.scripts]
actions_helper = 'actions_helper.main:cli'

[tool.coverage.run]
branch = true
source = ["."]
omit = ["tests/*.py"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
